seperate components into their own seperate files
