+++ apps/nectar/web/static/js/components/stacked_alerts.js
@@ -1,0 +1,13 @@
+import React from "react";
+import DismissableAlert from "./dismissable_alert";
+
+export default class StackedAlerts extends React.Component {
+  render() {
+    const alerts = this.props.alerts.map( (alert, idx) => <DismissableAlert message={alert.message} key={idx} index={idx} clear={this.props.clear}/>);
+    return (
+      <div>
+        {alerts}
+      </div>
+    );
+  }
+}
